File Manager.py

+++ File Manager.py
@@ -1,8 +1,42 @@
-#Generate a csv listing all files within directory incl. subdirectories
-##Rename List.csv to contain {File path} & {New file name} column headers
+#delete_fileless_branches() = deletes earliest ancestor directory containing zero files within itself incl. subfolders
+#list_files() = Generate a csv listing all files within directory incl. subdirectories
+#delete()/rename() = to complete their actions via their respective csv lists
+    ##Rename List.csv first row to have {File path} & {New file name} column headers
+    ##Delete List.csv first row to have {File path} column header
+#unique_output_check() = checks outputs filepath (incl. file name) are unique
+#unique_sources_check() = checks sources filepath (incl. file name) are unique
+
 import os
 import csv
 
+def delete_fileless_branches():
+    path = r"C:\Users\Dan\Desktop\New Folder"
+    empty = []
+    #identify empty folders and add folderpath to list
+    for root, dirs, files in os.walk(path):
+        if not len(dirs) and not len(files):
+            empty.append(root)
+
+    #for each item in empty list
+    #find earliest ancestor containing zero files within itself incl. subfolders
+    #then add folderpath to second list
+    fileless_branch = []
+    for empty_dir in empty:
+        parent_dir = empty_dir[:empty_dir.rfind("\\")]
+        while sum([len(files) for r, d, files in os.walk(parent_dir)]) ==0:
+            empty_dir = parent_dir
+            parent_dir = empty_dir[:empty_dir.rfind("\\")]
+            if parent_dir == path:
+                break
+
+        if empty_dir not in fileless_branch:
+            fileless_branch.append(empty_dir)
+
+    for i in empty:
+        print("empty: ", i)
+    for i in fileless_branch:
+        print("fileless_branch: ", i)
+    #Functionality to delete folderpaths in fileless_branch to be added
 
 def list_files():
     count = 0
@@ -23,11 +57,48 @@ def list_files():
     file.close()
     print("Done: ", count, " files have been listed.")
 
-fpath = r"C:\Users\Dan\Desktop\Test\Rename List.csv"
+#fpath = r"C:\Users\Dan\Desktop\Test\Rename List.csv"
+#fpath = r"C:\Users\Dan\Desktop\Test\Delete List.csv"
 
-def rename():
+
+def completion_message():
+    if len(missing) > 0:
+        print("[PARTIALLY COMPLETED] " + str(count) + " file(s) " + mode)
+        print("[SKIPPED FILE(S)] ", len(missing))
+        for i in missing:
+            print(i)
+    else:
+        print("[SUCCESSFULLY COMPLETED] " + str(count) + " file(s) " + mode)
+
+#delete() & rename() to be merged
+def delete():
+    global count
+    global missing
+    global mode
     count = 0
     missing = []
+    mode = "deleted"
+    #with open(r"C:\Users\Dan\Desktop\Test\Rename List.csv") as file_obj:
+    with open(fpath, "r", newline="") as file_obj:
+        reader_obj = csv.reader(file_obj)
+        next(reader_obj, None)
+        for row in reader_obj:
+            path = str(row[0])
+            if os.path.exists(path):
+                os.remove(path)
+                count += 1
+            else:
+                missing.append(path)
+
+    completion_message()
+
+def rename():
+    global count
+    global missing
+    global mode
+    count = 0
+    missing = []
+    mode = "renamed"
     #with open(r"C:\Users\Dan\Desktop\Test\Rename List.csv") as file_obj:
     with open(fpath, "r", newline="") as file_obj:
         reader_obj = csv.reader(file_obj)
@@ -37,21 +108,15 @@ def rename():
             dir = path[:path.rfind("\\")+1]
             #old_name = path[path.rfind("\\")+1:]
             new_name = dir + str(row[1])
-            try:
+            if os.path.exists(path):
                 os.rename(path, new_name)
                 count += 1
-            except:
+            else:
                 missing.append(path)
 
-    if len(missing) > 0:
-        print("[PARTIALLY COMPLETED] " + str(count) + " file(s) renamed")
-        print("[SKIPPED FILE(S)] ", len(missing))
-        for i in missing:
-            print(i)
-    else:
-        print("[SUCCESSFULLY COMPLETED] " + str(count) + " file(s) renamed")
+    completion_message()
 
-def repeated_output_check():
+def unique_output_check():
     #global file_obj
     outputs = []
     duplicate_outputs = []
@@ -77,7 +142,7 @@ def repeated_output_check():
     else:
         print("No output repetitions")
 
-def repeated_sources_check():
+def unique_sources_check():
     #global file_obj
     sources = []
     duplicate_sources = []
@@ -100,6 +165,11 @@ def repeated_sources_check():
     else:
         print("No source repetitions")
 
-repeated_sources_check()
-repeated_output_check()
-rename()
+def batch_rename():
+    unique_sources_check()
+    unique_output_check()
+    rename()
+
+def batch_delete():
+    unique_sources_check()
+    delete()
